requirements.txt
langgraph>=0.6.3
langchain>=0.3.27
langchain-core>=0.3.72
langchain-groq>=0.3.7
playwright>=1.40.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0